refactor(MoveRoute): move from command to event package
- delete command package
